requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
torch
torchvision
onnx
onnxruntime-gpu
transformers
einops
safetensors
timm
psd-tools
opencv-python